requirements.txt
python-telegram-bot==21.0.1
asyncpg>=0.29.0
redis>=5.0.0
python-dotenv>=1.0.0
uvloop>=0.19.0; sys_platform != 'win32'
